requirements.txt
streamlit
snowflake-connector-python[pandas]==2.8.0
seaborn
matplotlib
